move `ParserConfig` to separate file
